tav_opensource/backend/requirements.txt
# TAV Engine Requirements - Optimized for Deployment
# Only includes packages actually used by the codebase

# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
prometheus-client>=0.19.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Production reliability (Phase 1 improvements)
circuitbreaker>=1.4.0

# Environment and configuration
python-dotenv>=1.0.0

# CLI tools
click>=8.1.0

# Data processing and utilities
pyyaml>=6.0.0
requests>=2.31.0
pymupdf>=1.16.0
reportlab>=3.6.0

# Security and encryption
cryptography>=41.0.0

# JSON schema validation
jsonschema>=4.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
pydantic[email]
asyncpg>=0.29.0


# Date/time handling
python-dateutil>=2.8.0

# Async suppo rt
aiofiles>=23.0.0
aiohttp>=3.9.0
httpx>=0.24.0
httpx_aiohttp
redis>=4.6.0

# Machine Learning and AI (Core)
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0

# AI API Clients
openai>=1.0.0
anthropic>=0.25.0

# LangChain for multi-provider agent support
langchain==0.3.27
langchain-anthropic==0.3.19
langchain-community==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-mistralai==0.2.11
langchain-ollama==0.3.7
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.14
language_data==1.3.0
lazy_loader==0.4

# Hugging Face ecosystem
huggingface-hub>=0.20.0
huggingface_hub[hf_xet]
datasets>=2.14.0
accelerate>=0.21.0
sentencepiece>=0.1.99

# Computer Vision and Image Processing
Pillow>=10.0.0
matplotlib>=3.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Natural Language Processing
flair>=0.13.0
spacy>=3.7.0

# Communication APIs
twilio>=8.10.0

librosa>=0.10.0
litellm>=1.76.1

# Optional: Development tools (comment out for production)
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
# black>=23.0.0
# flake8>=6.0.0
# isort>=5.12.0 
ruff>=0.1.9